+++ src/test/java/seedu/address/storage/JsonAddressBookStorageTest.java
@@ -3,10 +3,10 @@ package seedu.address.storage;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static seedu.address.testutil.Assert.assertThrows;
+import static seedu.address.testutil.TypicalAddressBook.getTypicalAddressBook;
 import static seedu.address.testutil.TypicalMembers.ALICE;
 import static seedu.address.testutil.TypicalMembers.HOON;
 import static seedu.address.testutil.TypicalMembers.IDA;
-import static seedu.address.testutil.TypicalMembers.getTypicalAddressBook;
 
 import java.io.IOException;
 import java.nio.file.Path;
