+++ src/test/java/seedu/address/testutil/TypicalEvents.java
@@ -37,17 +37,6 @@ public class TypicalEvents {
 
     private TypicalEvents() {} // prevents instantiation
 
-    /**
-     * Returns an {@code AddressBook} with all the typical events.
-     */
-    public static AddressBook getTypicalAddressBook() {
-        AddressBook ab = new AddressBook();
-        for (Event event: getTypicalEvents()) {
-            ab.addEvent(event);
-        }
-        return ab;
-    }
-
     public static List<Event> getTypicalEvents() {
         return new ArrayList<>(Arrays.asList(AURORA_BOREALIS, BOXING_DAY, CHINESE_NEW_YEAR, DOG_DAY, ECHO_DAY,
                 FESTIVAL, GRAVITY_DISCOVERY_DAY));
