+++ src/test/java/seedu/address/testutil/TypicalMembers.java
@@ -59,17 +59,6 @@ public class TypicalMembers {
 
     private TypicalMembers() {} // prevents instantiation
 
-    /**
-     * Returns an {@code AddressBook} with all the typical members.
-     */
-    public static AddressBook getTypicalAddressBook() {
-        AddressBook ab = new AddressBook();
-        for (Member member : getTypicalMembers()) {
-            ab.addMember(member);
-        }
-        return ab;
-    }
-
     public static List<Member> getTypicalMembers() {
         return new ArrayList<>(Arrays.asList(ALICE, BENSON, CARL, DANIEL, ELLE, FIONA, GEORGE));
     }
